refactor(IN-314): 토큰 리프레시 후 유저 정보 반환
토큰 리프레시 성공 후 유저 정보를 반환하는 API 요청을 보내도록(/user/me)해 유저 상태를 업데이트 하도록 수정
