+++ src/app/layouts/SidebarStoreProvider.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { SidebarProvider } from '@/shared/ui/shadcn/sidebar'
-import { useSidebarStore } from '@/shared/api'
+import { useSidebarStore } from '@/widgets/layout/sidebar'
 
 export function SidebarStoreProvider({ children }: { children: React.ReactNode }) {
   const { open, setOpen } = useSidebarStore()
