+++ src/shared/api/index.ts
@@ -1,4 +1,4 @@
 export { axiosInstance } from './axiosInstance'
 export { useAuthStore } from './authStore'
-export { useSidebarStore } from './sidebarStore'
+export { fetchCurrentUser } from './userApi'
 export type { ApiResponse } from './types'
