+++ src/widgets/layout/sidebar/index.ts
@@ -1,1 +1,2 @@
 export { AppSidebar } from './ui/AppSidebar'
+export { useSidebarStore } from './model/sidebarStore'
